% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exitpoll.R
\name{empty_exitpoll_tibble}
\alias{empty_exitpoll_tibble}
\title{Return an empty tibble with the correct exitpoll column structure}
\usage{
empty_exitpoll_tibble()
}
\description{
Return an empty tibble with the correct exitpoll column structure
}
\keyword{internal}
