% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objectives.R
\name{empty_objective_results}
\alias{empty_objective_results}
\title{Return empty tibbles with correct column structure}
\usage{
empty_objective_results(slice_by, include_steps)
}
\description{
Return empty tibbles with correct column structure
}
\keyword{internal}
