% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objectives.R
\name{fetch_all_step_results}
\alias{fetch_all_step_results}
\title{Fetch step results for multiple objective versions}
\usage{
fetch_all_step_results(project_id, version_ids, session_filters, metadata)
}
\description{
Fetch step results for multiple objective versions
}
\keyword{internal}
