% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exitpoll.R
\name{fetch_exitpoll_hook_data}
\alias{fetch_exitpoll_hook_data}
\title{Fetch and parse all data for a single hook}
\usage{
fetch_exitpoll_hook_data(project_id, hook, version, session_filters)
}
\description{
Fetch and parse all data for a single hook
}
\keyword{internal}
