% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exitpoll.R
\name{fetch_exitpoll_hooks}
\alias{fetch_exitpoll_hooks}
\title{Fetch available exit poll hook names for a project}
\usage{
fetch_exitpoll_hooks(project_id)
}
\description{
Fetch available exit poll hook names for a project
}
\keyword{internal}
