% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exitpoll.R
\name{fetch_exitpoll_metadata}
\alias{fetch_exitpoll_metadata}
\title{Fetch and parse question set metadata for a specific hook}
\usage{
fetch_exitpoll_metadata(project_id, hook)
}
\description{
Fetch and parse question set metadata for a specific hook
}
\keyword{internal}
