% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objectives.R
\name{fetch_objectives_metadata}
\alias{fetch_objectives_metadata}
\title{Fetch and parse objectives metadata for a project}
\usage{
fetch_objectives_metadata(project_id)
}
\description{
Fetch and parse objectives metadata for a project
}
\keyword{internal}
