% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exitpoll.R
\name{fetch_response_counts}
\alias{fetch_response_counts}
\title{Fetch response counts for a single version}
\usage{
fetch_response_counts(project_id, hook, version, session_filters)
}
\description{
Fetch response counts for a single version
}
\keyword{internal}
