% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exitpoll.R
\name{map_value_label}
\alias{map_value_label}
\title{Map a raw response value to a human-readable label}
\usage{
map_value_label(value, question_type, question_meta)
}
\description{
Map a raw response value to a human-readable label
}
\keyword{internal}
