% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exitpoll.R
\name{parse_exitpoll_responses}
\alias{parse_exitpoll_responses}
\title{Parse response count arrays into a tidy tibble}
\usage{
parse_exitpoll_responses(raw_counts, version_num, hook, questions_meta)
}
\description{
Parse response count arrays into a tidy tibble
}
\keyword{internal}
