% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objectives.R
\name{parse_objective_results}
\alias{parse_objective_results}
\title{Parse objective results response and join with metadata}
\usage{
parse_objective_results(raw_results, slice_by, metadata)
}
\description{
Parse objective results response and join with metadata
}
\keyword{internal}
